Deep_Image_Prior/requirements.txt
matplotlib==3.5.2
numpy==1.24.4
Pillow==10.0.0
torch==2.0.0
tqdm==4.64.1